--- src/main/java/Actions/Action.java
@@ -1,8 +1,0 @@
-package Actions;
-
-import UI.Duke;
-
-
-public interface Action {
-    public void execute(Duke bot);
-}
