--- src/main/java/Actions/Bye.java
@@ -1,10 +1,0 @@
-package Actions;
-
-import UI.Duke;
-
-public class Bye implements Action {
-    @Override
-    public void execute(Duke bot) {
-        bot.exit();
-    }
-}
