--- src/main/java/Actions/Greet.java
@@ -1,10 +1,0 @@
-package Actions;
-
-import UI.Duke;
-
-public class Greet implements Action {
-    @Override
-    public void execute(Duke bot) {
-        System.out.println("Hello! I'm " + bot.getName() + "\nWhat can I do for you?\n");
-    }
-}
