--- src/main/java/Actions/InvalidAction.java
@@ -1,13 +1,0 @@
-package Actions;
-
-import UI.Duke;
-
-public class InvalidAction implements Action {
-    public InvalidAction() {
-    }
-
-    @Override
-    public void execute(Duke bot) {
-        System.out.println("Command does not exist");
-    }
-}
