--- src/main/java/ExceptionHandling/DukeException.java
@@ -1,7 +1,0 @@
-package ExceptionHandling;
-
-public class DukeException extends Exception{
-    public DukeException(String message) {
-        super(message);
-    }
-}
