--- src/main/java/UI/Main.java
@@ -1,4 +1,0 @@
-package UI;
-
-public class Main {
-}
