+++ src/main/java/actions/Action.java
@@ -1,0 +1,8 @@
+package actions;
+
+import ui.Duke;
+
+
+public interface Action {
+    public void execute(Duke bot);
+}
