+++ src/main/java/actions/Bye.java
@@ -1,0 +1,10 @@
+package actions;
+
+import ui.Duke;
+
+public class Bye implements Action {
+    @Override
+    public void execute(Duke bot) {
+        bot.exit();
+    }
+}
