+++ src/main/java/exceptionhandling/DukeException.java
@@ -1,0 +1,7 @@
+package exceptionhandling;
+
+public class DukeException extends Exception{
+    public DukeException(String message) {
+        super(message);
+    }
+}
