+++ src/main/java/ui/Main.java
@@ -1,0 +1,4 @@
+package ui;
+
+public class Main {
+}
